python-decompresser/requirements.txt
lz4
cffi
python-snappy